api-engine/requirements.txt
allure-combine==1.0.11
allure-pytest==2.13.5
allure-python-commons==2.13.5
async-timeout==5.0.1
attrs==23.2.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==8.7.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonpath==0.82.2
MarkupSafe==2.1.5
numpy==1.26.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
pika==1.3.2
pluggy==1.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.10.1
PyMySQL==1.1.0
PySocks==1.7.1
pytest==8.1.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
pyyaml-include==1.3.1
redis==5.2.1
requests==2.31.0
selenium==4.19.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2025.2
urllib3==2.2.1
Werkzeug==3.1.3
wsproto==1.2.0
xmindparser==1.0.9
zipp==3.23.0
python-dotenv>=1.0.0
